Reorganize packages
Main package is now called `malusers`.
New directory, `cmd`, was created, with directories/packages
`api` and `scraper` inside. I presume that `web` will be added there
at later date.
