+++ cmd/api/api.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/modrzew/malusers"
+)
+
+func main() {
+	user := &malusers.User{Username: "abc"}
+	fmt.Println(user)
+}
