+++ config.go
@@ -1,8 +1,9 @@
-package main
+package malusers
 
 import (
 	"encoding/json"
 	"os"
+	"path"
 )
 
 // Configuration holds database settings
@@ -12,7 +13,7 @@ type Configuration struct {
 	Database      string
 	Username      string
 	Password      string
-	SSLMode       string
+	SslMode       string
 	MaxConcurrent int
 }
 
@@ -24,10 +25,17 @@ func ReadConfig() *Configuration {
 	if read {
 		return &configuration
 	}
-	file, _ := os.Open("config.json")
+	dir, err := os.Getwd()
+	if err != nil {
+		panic(err)
+	}
+	file, err := os.Open(path.Join(dir, "config.json"))
+	if err != nil {
+		panic(err)
+	}
 	decoder := json.NewDecoder(file)
 	configuration = Configuration{}
-	err := decoder.Decode(&configuration)
+	err = decoder.Decode(&configuration)
 	if err != nil {
 		panic(err)
 	}
