+++ fetch.go
@@ -1,4 +1,4 @@
-package main
+package malusers
 
 import (
 	"fmt"
