+++ models.go
@@ -1,4 +1,4 @@
-package main
+package malusers
 
 import (
 	"github.com/jinzhu/gorm"
